ResultsView Clean Architecture Changes + University Entity/Factory Major Changes

+++ src/main/java/entity/UniversityFactory.java
@@ -1,6 +1,6 @@
 package entity;
 
 public interface UniversityFactory {
-    CommonUniversity create(Integer schoolID, String schoolName, String state, String city, Double admissionRate, Integer averageInStateTuition, Integer averageOutOfStateTuition, Double averageSATScore, Double averageACTScore, String url);
+    CommonUniversity create(Integer schoolID, String schoolName, String state, String city, Double admissionRate, Integer averageInStateTuition, Integer averageOutOfStateTuition, Integer averageSATScore, Integer averageACTScore, String url);
 
 }
